pyproject.toml
[tool.poetry]
name = "conan-cmake-cpp-project-tools"
version = "0.1.0"
description = ""
authors = ["CD Clark III <clifton.clark@gmail.com>"]
readme = "README.md"
packages = [{include = "conan_cmake_cpp_project_tools"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.7.0"
fspathtree = "^0.5"
pyyaml = "^6.0"
rich = "^12.6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ccc = "conan_cmake_cpp_project_tools.cli:app"
